#120 - NOSOUND as seperate option
