+++ include/pntr_app_cli.h
@@ -314,26 +314,6 @@ void pntr_app_platform_close(pntr_app* app) {
     app->platform = NULL;
 }
 
-pntr_sound* pntr_load_sound_from_memory(pntr_app_sound_type type, unsigned char* data, unsigned int dataSize) {
-    (void)type;
-    (void)data;
-    (void)dataSize;
-    return NULL;
-}
-
-void pntr_unload_sound(pntr_sound* sound) {
-    (void)sound;
-}
-
-void pntr_play_sound(pntr_sound* sound, bool loop) {
-    (void)sound;
-    (void)loop;
-}
-
-void pntr_stop_sound(pntr_sound* sound) {
-    (void)sound;
-}
-
 bool pntr_app_platform_update_delta_time(pntr_app* app) {
     // TODO: Make CLI delta time get the actual delta time.
     if (app->fps <= 0) {
